requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.3
attrs==23.1.0
autopep8==1.7.0
billiard==4.1.0
celery==5.3.4
certifi==2023.5.7
charset-normalizer==2.1.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.0.9
django-cors-headers==4.0.0
django-environ==0.9.0
django-filter==22.1
django-redis==5.4.0
django-safedelete==1.3.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.3.0
gunicorn==20.1.0
httmock==1.4.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
kombu==5.3.2
more-itertools==10.1.0
packaging==23.2
pluggy==0.13.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
py==1.11.0
pycodestyle==2.9.1
PyJWT==2.8.0
pytest==6.0.1
pytest-django==3.9.0
pytest-sugar==0.9.4
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
redis==5.0.1
requests==2.28.1
sentry-sdk==1.22.2
six==1.16.0
sqlparse==0.4.2
termcolor==2.3.0
toml==0.10.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.15
validators==0.22.0
vine==5.0.0
wcwidth==0.2.8
whitenoise==6.1.0
